image_generator/requirements.txt
black
fastapi
uvicorn[standard]
docker
Pillow
